refactor(pr-mode): split oversized source files under CI 1000-line limit (Refs #20)
CI enforces a hard 1000-line cap per .rs file (Source file length checks in
.github/workflows/ci.yml), a gate not run by local make ci-check. Eight files
added by the PR-mode work exceeded that cap. Split each by extracting cohesive
test modules and selector/conversion helpers into sibling files, registered via
#[cfg(test)] #[path] and module declarations. No behavior change.

- app_input/mod.rs        -> prs_orchestration.rs (dispatch helpers)
- app_input/prs.rs        -> prs_key_tests.rs (key-routing tests)
- app_input/prs_integration_tests.rs -> prs_integration_tests_lifecycle.rs
- github/tests.rs         -> tests_pr.rs, tests_pr_detail.rs (PR fixtures/tests)
- layout.rs               -> layout_tests.rs
- messages.rs             -> messages/event_conversion.rs (From impls)
- state/mod.rs            -> selectors.rs (AppState selectors)
- state/types.rs          -> pr_types.rs (PR domain state types)

Also fixed two clippy::map_unwrap_or sites and trimmed stale imports surfaced
by the extraction. Full suite: 756 tests pass; all clippy/complexity/allow/
architecture gates green; every src file now under the 1000-line cap.
